+++ packages/monte_carlo/active_region/source/src/MonteCarlo_ParticleSourceComponent.hpp
@@ -46,12 +46,12 @@ public:
   typedef Geometry::Model::CellIdSet CellIdSet;
 
   //! Constructor
-  ParticleSourceComponent( const size_t id,
+  ParticleSourceComponent( const uint32_t id,
                            const double selection_weight,
                            const std::shared_ptr<const Geometry::Model>& model );
 
   //! Constructor (with rejection cells)
-  ParticleSourceComponent( const size_t id,
+  ParticleSourceComponent( const uint32_t id,
                            const double selection_weight,
                            const CellIdSet& rejection_cells,
                            const std::shared_ptr<const Geometry::Model>& model );
@@ -81,7 +81,7 @@ public:
   double getSelectionWeight() const;
 
   //! Get the id of this source
-  size_t getId() const;
+  uint32_t getId() const;
 
   //! Return the number of sampling trials
   Counter getNumberOfTrials() const;
@@ -215,7 +215,7 @@ private:
   friend class boost::serialization::access;
 
   // The component id
-  UniqueIdManager<ParticleSourceComponent,size_t> d_id;
+  UniqueIdManager<ParticleSourceComponent,uint32_t> d_id;
 
   // The component selection weight
   double d_selection_weight;
