+++ packages/monte_carlo/active_region/source/src/MonteCarlo_StandardAdjointParticleSourceComponent.hpp
@@ -40,7 +40,7 @@ public:
 
   //! Constructor
   StandardAdjointParticleSourceComponent(
-    const size_t id,
+    const uint32_t id,
     const double selection_weight,
     const std::shared_ptr<const Geometry::Model>& model,
     const std::shared_ptr<const ParticleDistribution>& particle_distribution,
@@ -48,7 +48,7 @@ public:
 
   //! Constructor (with rejection cells )
   StandardAdjointParticleSourceComponent(
-    const size_t id,
+    const uint32_t id,
     const double selection_weight,
     const std::vector<Geometry::Model::EntityId>& rejection_cells,
     const std::shared_ptr<const Geometry::Model>& model,
@@ -57,7 +57,7 @@ public:
 
   //! Constructor (with rejection cells )
   StandardAdjointParticleSourceComponent(
-    const size_t id,
+    const uint32_t id,
     const double selection_weight,
     const CellIdSet& rejection_cells,
     const std::shared_ptr<const Geometry::Model>& model,
@@ -66,14 +66,14 @@ public:
 
   //! Constructor
   StandardAdjointParticleSourceComponent(
-    const size_t id,
+    const uint32_t id,
     const double selection_weight,
     const std::shared_ptr<const FilledGeometryModel>& model,
     const std::shared_ptr<const ParticleDistribution>& particle_distribution );
 
   //! Constructor (with rejection cells )
   StandardAdjointParticleSourceComponent(
-    const size_t id,
+    const uint32_t id,
     const double selection_weight,
     const std::vector<Geometry::Model::EntityId>& rejection_cells,
     const std::shared_ptr<const FilledGeometryModel>& model,
@@ -81,7 +81,7 @@ public:
 
   //! Constructor (with rejection cells )
   StandardAdjointParticleSourceComponent(
-    const size_t id,
+    const uint32_t id,
     const double selection_weight,
     const CellIdSet& rejection_cells,
     const std::shared_ptr<const FilledGeometryModel>& model,
